+++ resources/views/openapi/group.blade.php
@@ -1,0 +1,7 @@
+@if ($group->description !== null)
+    <div class="vellum-api-group-description">{!! $markdown($group->description) !!}</div>
+@endif
+
+@foreach ($group->operations as $operation)
+    @include('vellum::openapi.operation', ['operation' => $operation, 'markdown' => $markdown])
+@endforeach